Renamed SortComponent to ControlBar
